clean up namespaces and syntax in program code

+++ implement/pine/Platform/WebService/StartupAdminInterface.cs
@@ -1036,7 +1036,7 @@ public class StartupAdminInterface
 
                 report = report with
                 {
-                    storeReductionTimeSpentMilli = (int)0,
+                    storeReductionTimeSpentMilli = 0,
                 };
 
                 startPublicApp();
